requirements.txt
pandas
numpy
PYAML
matplotlib
seaborn
scikit-learn
statsmodels